Update installer, uninstaller, bashrc, and gitconfig

+++ uninstall.sh
@@ -5,7 +5,7 @@ dotfiles=(".tmux.conf" ".zshrc" ".conkyrc" ".bashrc" ".Xresources" ".gitconfig" 
 for file in "${dotfiles[@]}"; do
 	if [ -f "$HOME/$file" ]; then
 		echo "Removing ${HOME}/${file}"
-		rm $HOME/$file
+		rm "$HOME/$file"
 	fi
 done
 
@@ -14,6 +14,6 @@ directories=(".mpd" ".ncmpcpp" "bin" ".zsh")
 for directory in "${directories[@]}"; do
 	if [ -d "$HOME/$directory" ]; then
 		echo "Removing ${HOME}/${directory}"
-		rm -r $HOME/$directory
+		rm -r "${HOME:?}/${directory}"
 	fi
 done
